backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.14.0
anthropic==0.42.0
openai>=1.30.0
httpx==0.28.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
apscheduler==3.10.4
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.159.0
dnspython==2.7.0
beautifulsoup4==4.12.3
duckduckgo-search==7.5.1
